+++ lib/dao/diary_dao.dart
@@ -5,6 +5,7 @@ import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:woo_yeon_hi/model/enums.dart';
+import 'package:woo_yeon_hi/utils.dart';
 
 import '../model/diary_model.dart';
 
@@ -38,41 +39,48 @@ Future<void> saveDiary(Diary diary) async {
   });
 }
 
-Future<List<Map<String, dynamic>>> getDiaryData(int? user_idx, int filter_editor, int filter_sort, String filter_start, String filter_end) async {
+Future<List<Map<String, dynamic>>> getDiaryData(
+    int? user_idx,
+    int filter_editor,
+    int filter_sort,
+    String filter_start,
+    String filter_end) async {
   List<Map<String, dynamic>> results = [];
 
-  Query<Map<String, dynamic>> query = FirebaseFirestore.instance.collection('DiaryData');
-  
-  if(user_idx != null){
+  Query<Map<String, dynamic>> query =
+      FirebaseFirestore.instance.collection('DiaryData');
+
+  if (user_idx != null) {
     // 내가 쓴 일기
-    if(filter_editor == DiaryEditorState.EDITOR_USER.type){
+    if (filter_editor == DiaryEditorState.EDITOR_USER.type) {
       query = query.where('diary_user_idx', isEqualTo: user_idx);
     }
     // 상대방이 쓴 일기
-    else if(filter_editor == DiaryEditorState.EDITOR_LOVER.type){
+    else if (filter_editor == DiaryEditorState.EDITOR_LOVER.type) {
       query = query.where('diary_user_idx', isEqualTo: user_idx);
     }
   }
 
   // 기간 조회
-  if(filter_start.isNotEmpty && filter_end.isNotEmpty){
-    query = query.where('diary_date', isGreaterThanOrEqualTo: filter_start)
+  if (filter_start.isNotEmpty && filter_end.isNotEmpty) {
+    query = query
+        .where('diary_date', isGreaterThanOrEqualTo: filter_start)
         .where('diary_date', isLessThanOrEqualTo: filter_end);
-  }else if(filter_start.isNotEmpty && filter_end.isEmpty){
+  } else if (filter_start.isNotEmpty && filter_end.isEmpty) {
     query = query.where('diary_date', isGreaterThanOrEqualTo: filter_start);
-  }else if (filter_start.isEmpty && filter_end.isNotEmpty){
+  } else if (filter_start.isEmpty && filter_end.isNotEmpty) {
     query = query.where('diary_date', isLessThanOrEqualTo: filter_end);
   }
 
   // 정렬 조건
-  if(filter_sort == DiarySortState.SORT_ASC.type){
-    query = query.orderBy('diary_date', descending: false);
-  }else{
-    query = query.orderBy('diary_date', descending: true);
+  if (filter_sort == DiarySortState.SORT_ASC.type) {
+    query = query.orderBy('diary_idx', descending: false);
+  } else {
+    query = query.orderBy('diary_idx', descending: true);
   }
 
   var querySnapShot = await query.get();
-  for(var doc in querySnapShot.docs){
+  for (var doc in querySnapShot.docs) {
     results.add(doc.data());
   }
 
@@ -80,20 +88,39 @@ Future<List<Map<String, dynamic>>> getDiaryData(int? user_idx, int filter_editor
 }
 
 Future<void> uploadDiaryImage(XFile imageFile, String imageName) async {
-  await FirebaseStorage.instance.ref('image/diary/$imageName').putFile(File(imageFile.path));
+  await FirebaseStorage.instance
+      .ref('image/diary/$imageName')
+      .putFile(File(imageFile.path));
 }
 
 Future<Image> getDiaryImagePath(String path) async {
-  var imageURL = await FirebaseStorage.instance.ref('image/diary/$path').getDownloadURL();
-  var image = Image.network(imageURL, fit: BoxFit.cover,);
+  var imageURL =
+      await FirebaseStorage.instance.ref('image/diary/$path').getDownloadURL();
+  var image = Image.network(
+    imageURL,
+    fit: BoxFit.cover,
+  );
   return image;
 }
 
 Future<void> readDiary(Diary diary) async {
-  var querySnapshot = await FirebaseFirestore.instance.collection('DiaryData')
-      .where('diary_idx', isEqualTo: diary.diaryIdx).get();
+  var querySnapshot = await FirebaseFirestore.instance
+      .collection('DiaryData')
+      .where('diary_idx', isEqualTo: diary.diaryIdx)
+      .get();
   var document = querySnapshot.docs.first;
-  document.reference.update({
-    'diary_lover_check' : true
-  });
+  document.reference.update({'diary_lover_check': true});
+}
+
+Future<bool> isExistOnDate(DateTime date) async {
+  var stringDate = dateToString(date);
+  var querySnapshot = await FirebaseFirestore.instance
+      .collection('DiaryData')
+      .where('diary_date', isEqualTo: stringDate)
+      .get();
+  if (querySnapshot.docs.length > 0) {
+    return true;
+  } else {
+    return false;
+  }
 }
